Performance updates

See merge request m2mobi/foodscan!139

+++ Vittles/domain/src/main/java/com/example/domain/settings/model/PerformanceSetting.kt
@@ -6,8 +6,8 @@ package com.example.domain.settings.model
  * @property ms Milliseconds of the interval of the setting
  */
 
-enum class PerformanceSetting(var ms: Int) {
-    LOW(1000),
-    MEDIUM(500),
-    HIGH(200)
+enum class PerformanceSetting(var ms: Long) {
+    LOW(1000L),
+    MEDIUM(500L),
+    HIGH(200L)
 }
